Mahmoud Mustafa Oushi/document-qa-app/requirements.txt
streamlit
numpy
faiss-cpu
groq
sentence-transformers
scikit-learn
PyPDF2
pandas
ipywidgets
python-dotenv

